12.22

+++ src/server/ai/rag/kb_service.py
@@ -7,6 +7,7 @@ from urllib.parse import urljoin
 from src.enum import FileTypeEnum
 from src.configs import get_setting, logger
 from src.server.dto import ApiCommonResponseDTO
+from src.server.utils import TokenChecker
 from src.server.db.repository.ai_repository import create_kb_to_db, get_kb_list_from_db, delete_kb_from_db
 
 setting = get_setting()
@@ -14,12 +15,13 @@ kb_base_url = urljoin(setting.DIFY_SERVER_URL, 'datasets')
 kb_file_base_url = urljoin(setting.DIFY_SERVER_URL, 'datasets/')
 
 
-def create_kb(kb_name: str = Body(..., description="知识库名称"),
+def create_kb(token_checker: TokenChecker,
+              kb_name: str = Body(..., description="知识库名称"),
               kb_description: Optional[str] = Body(None, description="知识库描述")) -> ApiCommonResponseDTO:
     """创建 dify 知识库"""
     try:
-        # TODO user
-        user_id = "1"
+        if not (user_id := token_checker):
+            return ApiCommonResponseDTO(message="用户未登录!").model_dict()
         logger.info(kb_base_url)
         resp = requests.post(kb_base_url, headers={"Content-Type": "application/json",
                                                    "Authorization": f"Bearer {setting.DIFY_KB_SECRET_KEY}"},
@@ -41,12 +43,13 @@ def create_kb(kb_name: str = Body(..., description="知识库名称"),
         return ApiCommonResponseDTO(status=500, message="fail").model_dict()
 
 
-def get_kb_list(page: int = Query(1, description="页数"),
+def get_kb_list(token_checker: TokenChecker,
+                page: int = Query(1, description="页数"),
                 limit: int = Query(default=10, description="每页数据数")) -> ApiCommonResponseDTO:
     """获取 Dify 知识库列表  """
     try:
-        # TODO user
-        user_id = 1
+        if not (user_id := token_checker):
+            return ApiCommonResponseDTO(message="用户未登录!").model_dict()
         if data := get_kb_list_from_db(user_id=user_id, page_no=page, page_size=limit):
             return ApiCommonResponseDTO(status=200, data=data).model_dict()
         else:
@@ -57,9 +60,11 @@ def get_kb_list(page: int = Query(1, description="页数"),
         return ApiCommonResponseDTO(status=500, message="fail").model_dict()
 
 
-def delete_kb(kb_id: str = Body(..., description="kb_id")):
+def delete_kb(token_checker: TokenChecker, kb_id: str = Body(..., description="kb_id")):
     """删除dify 知识库"""
     try:
+        if not (user_id := token_checker):
+            return ApiCommonResponseDTO(message="用户未登录!").model_dict()
         delete_kb_from_db(kb_id=kb_id)
         return ApiCommonResponseDTO().model_dict()
     except BaseException as e:
@@ -68,10 +73,12 @@ def delete_kb(kb_id: str = Body(..., description="kb_id")):
         return ApiCommonResponseDTO(status=500, message="fail").model_dict()
 
 
-def upload_file_to_kb(kb_id: str = Body(..., description="kb_id"),
+def upload_file_to_kb(token_checker: TokenChecker,kb_id: str = Body(..., description="kb_id"),
                       file: UploadFile = File(..., description="上传图片")):
     """上传知识库文件"""
     try:
+        if not (user_id := token_checker):
+            return ApiCommonResponseDTO(message="用户未登录!").model_dict()
         file_upload_url = urljoin(kb_file_base_url, f'{kb_id}/document/create-by-file')
         logger.info(file_upload_url)
         resp = requests.post(file_upload_url, headers={"Authorization": f"Bearer {setting.DIFY_KB_SECRET_KEY}"},
@@ -93,9 +100,11 @@ def upload_file_to_kb(kb_id: str = Body(..., description="kb_id"),
         return ApiCommonResponseDTO(status=500, message="fail").model_dict()
 
 
-def upload_text_to_kb(kb_id: str = Body(..., description="kb_id")):
+def upload_text_to_kb(token_checker: TokenChecker,kb_id: str = Body(..., description="kb_id")):
     """上传知识库文本 """
     try:
+        if not (user_id := token_checker):
+            return ApiCommonResponseDTO(message="用户未登录!").model_dict()
         text_upload_url = urljoin(kb_file_base_url, f'/{kb_id}/document/create-by-file')
         resp = requests.post(text_upload_url, headers={"Content-Type": "application/json",
                                                        "Authorization": f"Bearer {setting.DIFY_KB_SECRET_KEY}"},
